Greedy_Search_ algorithm[C#]
A greedy algorithm is an algorithm that follows the problem solving
heuristic of making the locally optimal choice at each stage with the
hope of finding a global optimum. In many problems, a greedy strategy
does not in general produce an optimal solution, but nonetheless a
greedy heuristic may yield locally optimal solutions that approximate a
global optimal solution in a reasonable time.